WARNING: This version of DrFTPD+ is unstable test code, and you should probably NOT use it in a production environment.
An announcement will be made in the forums and this trunk will be copied to a +STABLE+ branch when this code is ready for use by the masses.
In the meantime every effort will be made to keep this code as bug-free as possible, but expect it to not even compile on occasion as we work on it.
Once the +STABLE+ branch has been created, most users will probably want to follow that instead of this trunk.