PLUS-trunk:  * The class used to render MLST/MLSD commands also handles VFS serialization (files.mlst).  This fix should keep them both in their appropriate time zones, sorry for the double-commit.
While one could argue that files.mlst should probably store datestamps in GMT instead of the local time zone, historically this has not been the case, and I think it is preferable to not wreak havoc on peoples existing installations.
 
'''again this fix works for me on my test server set to a non-GMT zone, but it is no more widely tested than that.  If it does not work for you please report it!'''