Introduce a custom EventBus implementation that allows us to publish events asynchronously, this extends the default ThreadSafeEventService and provides publishAsync() methods to match the existing publish() methods, as a result events can be published either synchronously or asynchronously as appropriate.
Events that are published asynchronously are placed in a queue where another thread picks them up and publishes them to the appropriate syncrhonous method, this allows the original publisher to return very quickly and gives us a performance boost in scenarios such as publishing events after an upload since control will return to the ftp control thread quicker, this means less pause between files. If necessary this could be extended to support configurably pooled publishing threads but as of now I'm unsure if this is actually necessary and also whether it could cause issues with events being published in an unexpected order.

As a result of this it was necessary to switch back to using GlobalContext to access the event service so that it could be returned with a type of the new asynchronous service, not doing this would have required runtime casts everywhere it was retrieved to cast it from the EventService interface.

All existing publish calls have been reviewed and the majority changed to asynchronous, if anyone feels any of these have been changed inappropriately or should have been changed but haven't then please let me know with reasoning, the main remaining publish events that are called synchronously are:
 * Plugin load/unload and config reload, here I felt it desirable to not return from the command until the action had really been completed, returning whilst the event was still being processed could lead to confusion.
 * Events fired off from user stat resets, here the published data gets modified after the publish call so publishing asynchronously could lead to the data being different by the time it is actually processed.