Fix another place missed when standardising capitalisation of plugin parameter names.
This highlighted a strange access issue in calling CommonPluginUtils.loadClass() from MasterPluginUtils.getLoadedExtensionObjects(), this call was throwing a runtime IllegalAccessException. Intuitively this seems odd as both of these classes are in the org.drftpd.util package and CommonPluginUtils.loadClass() is a protected method so should have visibility within the same package. I suspect the problem was that although they are in the same package they are loaded by different plugins and therefore by different classloaders, I didn't want to make CommonPluginUtils.loadClass() public as I don't want third party authors calling it, the simplest workaround for this was to make MasterPluginUtils extend CommonPluginUtils and this solved the access issue.

Given the above we should probably consider making sure we do not split the same package name across mutiple plugins.