Add some of the infrastructure for controlling min/max speed restrictions on file transfers. This provides a mechanism to set a min/max and passing of the values populated into the transfer code on the slave. An actual implementation of the rate limiting in the slave transfer code is not included in this changeset.
In order to pass min/max speeds a plugin should prehook the appropriate command (RETR or STOR) and add one or both of the DataConnectionHandler.MIN_XFER_SPEED and DataConnectionHandler.MAX_XFER_SPEED static Key objects to the CommandRequest keyed map with an appropriate long value. It is anticapted that the long value will be in bytes/second, a value of 0 (which is the default if a prehook does not set these) is interpreted as no limit imposed.

Also converted all core issuing of commands to slaves to use the array arguments, this removes one known failure point for commas in filenames/dirnames, however there may still be others present (this is untested).

NOTE: This changes both master and slave code along with the protocol classes, when upgrading to this changeset all slaves *MUST* be updated at the same time.